LayoutTests/platform/wpe/imported/w3c/web-platform-tests/xhr/formdata-blob-expected.txt


FAIL formdata with blob assert_equals: expected "\nkey=blob:text/x-value:5," but got "\nkey=blob:text/x-value:0,"
FAIL formdata with named blob assert_equals: expected "\nkey=blob.txt:text/x-value:5," but got "\nkey=blob.txt:text/x-value:0,"
PASS formdata.append() should throw if value is string and file name is given 
